Refactor Components
