+++ frontend/src/App.jsx
@@ -1,14 +1,12 @@
 import Home from "./pages/Home";
+import Dashboard from "./pages/Dashboard";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { lazy, Suspense } from "react";
+import Transfer from "./pages/Transfer";
+import Partnership from "./pages/Partnership";
 const SignUp = lazy(() => import("./pages/SignUp"));
 const Login = lazy(() => import("./pages/Login"));
-const Dashboard = lazy(() => import("./pages/Dashboard"));
 const Profile = lazy(() => import("./pages/Profile"));
-const Clients = lazy(() => import("./pages/Clients"));
-const Vendors = lazy(() => import("./pages/Vendors"));
-const Invoices = lazy(() => import("./pages/Invoices"));
-const Payment = lazy(() => import("./pages/Payment"));
 const Expenses = lazy(() => import("./pages/Expenses"));
 
 function App() {
@@ -21,12 +19,10 @@ function App() {
           <Route path="/signup" element={<SignUp />} />
           <Route path="/login" element={<Login />} />
           <Route path="/profile" element={<Profile />} />
-          <Route path="/clients" element={<Clients />} />
-          <Route path="/vendors" element={<Vendors />} />
-          <Route path="/invoices" element={<Invoices />} />
-          <Route path="/payment" element={<Payment />} />
           <Route path="/expenses" element={<Expenses />} />
           <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/transfer" element={<Transfer />} />
+          <Route path="/partnership" element={<Partnership />} />
         </Routes>
       </Suspense>
     </Router>
