+++ frontend/src/components/ExpensesPie.jsx
@@ -1,12 +1,17 @@
 import { Pie } from "react-chartjs-2";
 // eslint-disable-next-line no-unused-vars
-import { Chart as ChartJS } from "chart.js/auto";
+import { Chart as ChartJS, Legend, plugins } from "chart.js/auto";
 import PropTypes from "prop-types";
 
 const options = {
   responsive: true,
   // maintainAspectRatio: false,
-  cutout: 70,
+  cutout: 80,
+  plugins: {
+    legend: {
+      display: false,
+    },
+  },
 };
 
 const ExpensesPie = (props) => {
