+++ frontend/src/sass/pages/_home.scss
@@ -59,7 +59,7 @@
           font-weight: 600;
           margin-bottom: 1.2em;
         }
-        button {
+        a {
           padding: 10px 20px;
           background-color: $white;
           border: none;
