+++ frontend/src/sass/pages/_index.scss
@@ -7,3 +7,5 @@
 @forward "./invoices";
 @forward "./payment";
 @forward "./expenses";
+@forward "./partnership";
+@forward "./transfer";
