+++ frontend/src/sass/pages/_invoices.scss
@@ -5,10 +5,8 @@
   .title {
     font-family: $primary-font-stack;
     font-size: 2em;
-    margin-bottom: 1rem;
-  }
-  .description {
-    margin-bottom: 2rem;
+    margin-block: 2em 1rem;
+    margin-left: 7em;
   }
   .table-container {
     .add-invoice {
