+++ frontend/src/sass/pages/_partnership.scss
@@ -1,0 +1,10 @@
+@use "../abstracts" as *;
+
+#partnership {
+  h1.title {
+    margin: 1em;
+    font-family: $primary-font-stack;
+    font-size: 2em;
+    color: $blue;
+  }
+}
