+++ frontend/src/sass/pages/_payment.scss
@@ -5,7 +5,8 @@
   .title {
     font-family: $primary-font-stack;
     font-size: 2em;
-    margin-bottom: 1rem;
+    margin-block: 2em 1rem;
+    margin-left: 7em;
   }
   .description {
     margin-bottom: 2rem;
