+++ frontend/src/sass/pages/_profile.scss
@@ -1,7 +1,7 @@
 @use "../abstracts" as *;
 
 #profile {
-  font-size: 1rem;
+  font-size: 1.2rem;
   .container {
     .title {
       font-family: $primary-font-stack;
